Refactor: move function to CouponDetails repo.
